Update input files to match General Network Code.
In past versions of this code, the metric_input.xlsx files listed first path survival and then path probabilities for each season, this was different than the setup in the Generalized Network Code, making the two incompatible. This update fixes this problem by correctly ordering the metric_input files and changing the GR.R code to read in the data in the correct locations. This means the code is not backward compatible so all old input files must be reordered before using the latest version.